@@ -1,17 +1,64 @@
 package com.example.demo;
 
+import java.math.BigDecimal;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.ApplicationArguments;
+import org.springframework.boot.ApplicationRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+import com.example.demo.model.persistence.Item;
+import com.example.demo.model.persistence.User;
+import com.example.demo.model.persistence.repositories.ItemRepository;
+import com.example.demo.model.persistence.repositories.UserRepository;
 
 @EnableJpaRepositories("com.example.demo.model.persistence.repositories")
 @EntityScan("com.example.demo.model.persistence")
 @SpringBootApplication
-public class SareetaApplication {
+public class SareetaApplication extends SpringBootServletInitializer implements ApplicationRunner {
+
+	@Autowired
+	UserRepository userRepository;
+	
+    @Autowired
+    ItemRepository itemRepository;
+
+	@Autowired
+	PasswordEncoder passwordEncoder;
 
 	public static void main(String[] args) {
 		SpringApplication.run(SareetaApplication.class, args);
 	}
 
+	@Override
+	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+		return application.sources(SareetaApplication.class);
+	}
+
+	@Override
+	public void run(ApplicationArguments args) throws Exception {
+		User user = new User();
+		user.setUsername("admin");
+        user.setPassword(passwordEncoder.encode("admin"));
+        userRepository.save(user);
+        System.out.println(user);
+
+        Item item = new Item();
+        item.setName("Round Widget");
+        item.setPrice(new BigDecimal("2.99"));
+        item.setDescription("A widget that is round");
+        itemRepository.save(item);
+
+        Item item1 = new Item();
+        item1.setName("Square Widget");
+        item1.setPrice(new BigDecimal("1.99"));
+        item1.setDescription("A widget that is square");
+        itemRepository.save(item1);
+    }
 }
