@@ -4,14 +4,36 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class CreateUserRequest {
 
-	@JsonProperty
-	private String username;
+    @JsonProperty
+    private String username;
 
-	public String getUsername() {
-		return username;
-	}
+    @JsonProperty
+    private String password;
 
-	public void setUsername(String username) {
-		this.username = username;
-	}
+    @JsonProperty
+    private String rePassword;
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public String getRePassword() {
+        return rePassword;
+    }
+
+    public void setRePassword(String rePassword) {
+        this.rePassword = rePassword;
+    }
 }
